requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow>=10.0.0
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
numpy>=1.24.0
requests==2.31.0
python-dotenv==1.0.0